pyproject.toml
[project]
name = "kasflow"
version = "0.1.0"
description = "AI-powered personal expense tracker Telegram bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==22.3",
    "langgraph==0.5.4",
    "langchain-openai==0.3.28",
    "langchain-core==0.3.70",
    "trustcall==0.0.39",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "aiosqlite==0.21.0",
]

[dependency-groups]
dev = [
    "notebook>=7.4.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[project.scripts]
kasflow = "kasflow.main:run"

[build-system]
requires = ["uv_build>=0.8.0,<0.9.0"]
build-backend = "uv_build"
